+++ modules/ccv2/core/src/sap/commerce/toolset/ccv2/settings/state/CCv2ApplicationSettingsState.kt
@@ -24,21 +24,24 @@ import com.intellij.util.xmlb.annotations.Tag
 @Tag("HybrisApplicationSettings")
 data class CCv2ApplicationSettingsState(
     @JvmField @OptionTag val readTimeout: Int = 60,
+    @JvmField @OptionTag val authentication: CCv2Authentication = CCv2Authentication(),
     @JvmField val subscriptions: List<CCv2Subscription> = emptyList()
 ) {
     fun mutable() = Mutable(
-        readTimeout = readTimeout,
+        origin = this,
+        authentication = authentication.mutable(),
         subscriptions = subscriptions
             .map { it.mutable() }
             .toMutableList()
     )
 
     data class Mutable(
-        var readTimeout: Int,
+        private val origin: CCv2ApplicationSettingsState,
+        var authentication: CCv2Authentication.Mutable,
         var subscriptions: MutableList<CCv2Subscription.Mutable>
     ) {
         fun immutable() = CCv2ApplicationSettingsState(
-            readTimeout = readTimeout,
+            authentication = authentication.immutable(),
             subscriptions = subscriptions
                 .map { it.immutable() }
         )