+++ modules/project/core/src/sap/commerce/toolset/project/context/ProjectImportContext.kt
@@ -41,7 +41,6 @@ data class ProjectImportContext(
     val settings: ProjectImportSettings,
 
     val modulesFilesDirectory: Path? = null,
-    val ccv2Token: String? = null,
     val sourceCodePath: Path? = null,
     val sourceCodeFile: Path? = null,
     val projectIconFile: Path? = null,
@@ -116,7 +115,6 @@ data class ProjectImportContext(
 
         var project: Project? = null,
         var modulesFilesDirectory: Path? = null,
-        var ccv2Token: String? = null,
         var sourceCodePath: Path? = null,
         var sourceCodeFile: Path? = null,
         var projectIconFile: Path? = null,
@@ -161,7 +159,6 @@ data class ProjectImportContext(
             removeExternalModules = removeExternalModules,
 
             modulesFilesDirectory = modulesFilesDirectory,
-            ccv2Token = ccv2Token,
             sourceCodePath = sourceCodePath,
             sourceCodeFile = sourceCodeFile,
             projectIconFile = projectIconFile,